+++ pkg/googlesheets/utils.go
@@ -13,3 +13,17 @@ func findTimeField(frame *data.Frame) int {
 	}
 	return -1
 }
+
+func getExcelColumnName(columnNumber int) string {
+	dividend := columnNumber
+	columnName := ""
+	var modulo int
+
+	for dividend > 0 {
+		modulo = ((dividend - 1) % 26)
+		columnName = string(65+modulo) + columnName
+		dividend = ((dividend - modulo) / 26)
+	}
+
+	return columnName
+}
